google-adk-example/pyproject.toml
[project]
name = "google-adk-example"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "google-adk>=1.31.0",
    "uvicorn",
]
